tidy: fix remaining clang-tidy errors from arm_tidy CI
- readability-redundant-casting: drop `static_cast<double>` of the already-`double`
  `false_positive_rate` in BloomFilterHash (became redundant after the `long double` ->
  `double` change in this PR).
- readability-container-contains: convert `find() == end()`/`== npos` to `contains()` in
  KeeperStorage, BackupInfo, maskSensitiveQueryParameters and a backup test (code merged
  from master after the original report).
- readability-redundant-parentheses: remove redundant parens around the ternary operands
  in StorageMergeTree (the next redundant pair on a line already partially fixed).

https://s3.amazonaws.com/clickhouse-test-reports/PRs/108400/7f9b6822bc2b0420bb33cc91583a32e41a7563c1/build_arm_tidy/build_clickhouse/build_clickhouse.log
https://github.com/ClickHouse/ClickHouse/pull/108400

+++ src/Interpreters/BloomFilterHash.h
@@ -411,11 +411,11 @@ struct BloomFilterHash
 
         for (size_t bits_per_row = 1; bits_per_row < MAX_BITS_PER_ROW; ++bits_per_row)
         {
-            if (probability_lookup_table[bits_per_row][min_probability_index_each_bits[bits_per_row]] <= static_cast<double>(false_positive_rate))
+            if (probability_lookup_table[bits_per_row][min_probability_index_each_bits[bits_per_row]] <= false_positive_rate)
             {
                 size_t max_size_of_hash_functions = min_probability_index_each_bits[bits_per_row];
                 for (size_t size_of_hash_functions = max_size_of_hash_functions; size_of_hash_functions > 0; --size_of_hash_functions)
-                    if (probability_lookup_table[bits_per_row][size_of_hash_functions] > static_cast<double>(false_positive_rate))
+                    if (probability_lookup_table[bits_per_row][size_of_hash_functions] > false_positive_rate)
                         return std::pair<size_t, size_t>(bits_per_row, size_of_hash_functions + 1);
             }
         }
